+++ analytics/anvil-explorer-sheets/constants.py
@@ -9,12 +9,10 @@ END_DATE_PRIOR = "2024-12-31"
 PARENT_FOLDER_NAME = "January 2025"
 
 # The name of the spreadsheet with the report
-SHEET_NAME = "HCA Explorer"
-
-HCA_ID = "361323030"
-# Filter to exclude the Data Explorer
-HCA_BROWSER_ONLY_FILTER = {"filter": {"fieldName": "hostName", "stringFilter": {"matchType": "EXACT", "value": "explore.data.humancellatlas.org"}}}
-SECRET_NAME = "HCA_ANALYTICS_REPORTING_CLIENT_SECRET_PATH"
-ANALYTICS_START = "2024-02-01"
+SHEET_NAME = "AnVIL Explorer"
+ANVIL_EXPLORER_ID = "383267328"
+SECRET_NAME = 'ANVIL_ANALYTICS_REPORTING_CLIENT_SECRET_PATH'
+GA_PROPERTY_PORTAL = "368678391" # AnVIL Explorer - GA4
+ANALYTICS_START = "2024-01-01"
 
 OAUTH_PORT = 8082