+++ app/viewModelBuilders/common/contants.ts
@@ -6,5 +6,5 @@ export const DATE_TIME_FORMAT_OPTIONS: Intl.DateTimeFormatOptions = {
 };
 export const DATE_TIME_LOCALES = "en-US";
 export const FEATURE_FLAGS = {
-  VERBATIM: "verbatim",
+  VERBATIM: "verbatim", // Deprecated
 };