+++ site-config/anvil-cmg/dev/export/export.ts
@@ -3,7 +3,6 @@ import {
   ExportConfig,
 } from "@databiosphere/findable-ui/lib/config/entities";
 import * as C from "../../../../app/components";
-import { ExportMethod } from "../../../../app/components/Export/components/AnVILExplorer/ExportMethod/exportMethod";
 import * as V from "../../../../app/viewModelBuilders/azul/anvil-cmg/common/viewModelBuilders";
 import { ROUTE_EXPORT_TO_TERRA, ROUTE_MANIFEST_DOWNLOAD } from "./constants";
 import { mainColumn as exportMainColumn } from "./exportMainColumn";
@@ -73,9 +72,9 @@ export const exportConfig: ExportConfig = {
         {
           children: [
             {
-              component: ExportMethod,
+              component: C.ExportMethod,
               viewBuilder: V.buildExportMethodTerra,
-            } as ComponentConfig<typeof ExportMethod>,
+            } as ComponentConfig<typeof C.ExportMethod>,
             {
               component: C.ExportMethod,
               viewBuilder: V.buildExportMethodManifestDownload,